# FamilyClub.React/famiy.club.react/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "FamilyClub.WebAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/ClubMember/{id}": {
      "get": {
        "tags": [
          "ClubMember"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.ClubMember.ClubMemberReadDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.ClubMember.ClubMemberReadDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.ClubMember.ClubMemberReadDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ClubMember"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Order.OrderDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Order.OrderDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Order.OrderDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Order.OrderDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Order.OrderDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Order.OrderDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Order.OrderDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Order.OrderDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Order.OrderDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Order.OrderDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Order.OrderDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Order.OrderDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Orders/by-user/{userId}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Order.OrderDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Order.OrderDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Order.OrderDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Promotions": {
      "get": {
        "tags": [
          "Promotions"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Promotion.PromotionDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Promotion.PromotionDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Promotion.PromotionDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Promotions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Promotion.PromotionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Promotion.PromotionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Promotion.PromotionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Promotions/{id}": {
      "get": {
        "tags": [
          "Promotions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Promotion.PromotionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Promotion.PromotionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Promotion.PromotionDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Promotions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Promotion.PromotionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Promotion.PromotionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Promotion.PromotionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Promotions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/RolesClubMember": {
      "get": {
        "tags": [
          "RolesClubMember"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "RolesClubMember"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/RolesClubMember/{id}": {
      "get": {
        "tags": [
          "RolesClubMember"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "RolesClubMember"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "RolesClubMember"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/RolesClubMember/{roleName}/users": {
      "get": {
        "tags": [
          "RolesClubMember"
        ],
        "parameters": [
          {
            "name": "roleName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FamilyClub.BLL.DTOs.ClubMember.ClubMemberReadDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "avatarData": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "surname": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FamilyClub.BLL.DTOs.Order.OrderDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "orderDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "orderItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.OrderItem.OrderItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FamilyClub.BLL.DTOs.OrderItem.OrderItemDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "unitPrice": {
            "type": "number",
            "format": "double"
          },
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "orderId": {
            "type": "integer",
            "format": "int32"
          },
          "format": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FamilyClub.BLL.DTOs.Promotion.PromotionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "discountPercent": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}